+++ src/graphology/src/data.js
@@ -102,19 +102,11 @@ UndirectedNodeData.prototype.upgradeToMixed = function () {
  * @constructor
  * @param {boolean} undirected   - Whether the edge is undirected.
  * @param {string}  string       - The edge's key.
- * @param {boolean} generatedKey - Was its key generated?
  * @param {string}  source       - Source of the edge.
  * @param {string}  target       - Target of the edge.
  * @param {object}  attributes   - Edge's attributes.
  */
-export function EdgeData(
-  undirected,
-  key,
-  generatedKey,
-  source,
-  target,
-  attributes
-) {
+export function EdgeData(undirected, key, source, target, attributes) {
   // Attributes
   this.key = key;
   this.attributes = attributes;
@@ -123,7 +115,4 @@ export function EdgeData(
   // Extremities
   this.source = source;
   this.target = target;
-
-  // Was its key generated?
-  this.generatedKey = generatedKey;
 }